Split up transform into its own module. (#499)
* Added chain method to Transformation.
